requirements.txt
# fab-classic is a Fabric fork with py3 support, faithful to the py27 API
fab-classic==1.18.1
pyyaml==5.3.1
